More information for creating a PTR Setup
When trying to setup a server I only had a VPS at a local hosting provider. The DNS was setup through cloudflare. Since its the first time I set up the reverse pointer I where to set it up or how to verify it. This change should help others in the same situation by hinting that it has to do with the server host (as opposed to domain host) and that you can use `nslookup` to verify that it works.